Isolate websocket and http from each other on the server - Also fix http status display for the websocket
