+++ ecosystem.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    apps: [
+        {
+            name: "nt-http",
+            script: "yarn server-http"
+        },
+        {
+            name: "nt-ws",
+            script: "yarn server-ws"
+        }
+    ]
+}
