+++ nt-web-app/package.json
@@ -8,7 +8,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "serve": "ts-node server.ts",
+    "serve-http": "ts-node nt-webserver.ts",
+    "serve-websocket": "ts-node nt-websocket.ts",
     "serverInitOnce": "node create-jwt-secret.js",
     "serveDev": "ts-node-dev server.ts",
     "type-gen2": "pbjs -t static-module --ts --no-comments --no-delimited --out websocket/messages.ts websocket/messages.proto",
