+++ package.json
@@ -12,7 +12,8 @@
   "scripts": {
     "client": "yarn workspace nt-app electron:serve",
     "serverInitOnce": "yarn workspace nt-web-app serverInitOnce",
-    "server": "yarn workspace nt-web-app serve",
+    "server-http": "yarn workspace nt-web-app serve-http",
+    "server-ws": "yarn workspace nt-web-app serve-websocket",
     "serverDev": "yarn workspace nt-web-app serveDev",
     "buildClient": "yarn workspace nt-app electron:build",
     "regenerate-proto-experimental": "pbjs -t static-module --ts --no-comments --no-delimited --out websocket/messages.ts websocket/messages.proto",
@@ -26,6 +27,8 @@
   },
   "devDependencies": {
     "pbjs": "^0.0.14",
-    "prettier": "^2.8.8"
+    "prettier": "^2.8.8",
+    "ts-node": "^10.9.1",
+    "ts-node-dev": "^2.0.0"
   }
 }
